pyproject.toml
[project]
name = "HealthMonitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.136.1",
    "paho-mqtt>=2.1.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]
